957: Rename to PaymentController
