+++ app/javascript/controllers/index.ts
@@ -7,5 +7,5 @@ import { application } from "./application"
 import Gdrive__RootPickerController from "./gdrive/root_picker_controller"
 application.register("gdrive--root-picker", Gdrive__RootPickerController)
 
-import SubscriptionController from "./subscription_controller"
-application.register("subscription", SubscriptionController)
+import PaymentController from "./payment_controller"
+application.register("payment", PaymentController)
