+++ app/javascript/controllers/payment_controller.ts
